requirements.txt
Django==4.2.13
django-jazzmin==3.0.0
django-tinymce==4.0.0
pillow==10.3.0
